+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2026-08-10
+
+### Added
+
+- Added Caverns and Chasms Copper hammers and their oxidized variants (thanks, BurningSulphur!).
+
+### Fixed
+
+- Fixed server crash when stirring Mortar with Create installed.
+- Fixed mod compatibility hammers missing tags.
+
 ## [2.1.2] - 2026-08-09
 
 ### Fixed